Make q,k,v dtype all the same bfloat16 for FusedSDPA

+++ optimum/habana/transformers/models/mixtral/modeling_mixtral.py
@@ -351,7 +351,8 @@ class GaudiMixtralAttention(MixtralAttention):
 
         if FusedSDPA:
             if query_states.dtype != key_states.dtype:
-                query_states = query_states.type(key_states.dtype)
+                key_states = key_states.type(query_states.dtype)
+                value_states = value_states.type(query_states.dtype)
             # support long sequences exceeding 8192
             if not self.training and q_len == key_states.size(-2) and q_len > 8192:
                 htcore.mark_step()
